Added another decorator classes with refactoring -Till Chapter 6
