+++ src/Logic/AppServices/DisenrollCommand.cs
@@ -1,0 +1,60 @@
+﻿using CSharpFunctionalExtensions;
+using Logic.Students;
+using Logic.Utils;
+
+namespace Logic.AppServices
+{
+
+  #region DisenrollCommand
+  public sealed class DisenrollCommand : ICommand
+  {
+    public long Id { get; }
+    public int EnrollmentNumber { get; }
+    public string Comment { get; }
+
+    public DisenrollCommand(long id, int enrollmentNumber, string comment)
+    {
+      Id = id;
+      EnrollmentNumber = enrollmentNumber;
+      Comment = comment;
+    }
+
+    internal sealed class DisenrollCommandHandler : ICommandHandler<DisenrollCommand>
+    {
+      private readonly SessionFactory _sessionFactory;
+
+      public DisenrollCommandHandler(SessionFactory sessionFactory)
+      {
+        _sessionFactory = sessionFactory;
+      }
+
+      public Result Handle(DisenrollCommand command)
+      {
+        var unitOfWork = new UnitOfWork(_sessionFactory);
+        var studentRepository = new StudentRepository(unitOfWork);
+
+        Student student = studentRepository.GetById(command.Id);
+        if (student == null)
+          return Result.Fail($"No student found for Id {command.Id}");
+
+        if (string.IsNullOrWhiteSpace(command.Comment))
+          return Result.Fail("Disenrollment comment is required");
+
+        Enrollment enrollment = student.GetEnrollment(command.EnrollmentNumber);
+        if (enrollment == null)
+          return Result.Fail($"No enrollment found with number ' { command.EnrollmentNumber } '");
+
+        student.RemoveEnrollment(enrollment, command.Comment);
+
+        unitOfWork.Commit();
+
+        return Result.Ok();
+      }
+
+    }
+  }
+
+  
+  #endregion
+
+}
