--- src/Logic/Students/EditPersonalInfoCommand.cs
@@ -1,373 +1,0 @@
-﻿using CSharpFunctionalExtensions;
-using Logic.Utils;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Logic.Students
-{
-  #region EditPersonalInfoCommand
-  public interface ICommand
-  {
-
-  }
-
-  public interface ICommandHandler<TCommand>
-    where TCommand : ICommand
-  {
-    Result Handle(TCommand command);
-  }
-
-  public sealed class EditPersonalInfoCommand : ICommand
-  {
-    public long Id { get; }
-    public string Name { get; }
-    public string Email { get; }
-
-    public EditPersonalInfoCommand(long id, string name, string email)
-    {
-      Id = id;
-      Name = name;
-      Email = email;
-    }
-  }
-
-  public sealed class EditPersonalInfoCommandHandler : ICommandHandler<EditPersonalInfoCommand>
-  {
-    private readonly SessionFactory _sessionFactory;
-
-    public EditPersonalInfoCommandHandler(SessionFactory sessionFactory)
-    {
-      _sessionFactory = sessionFactory;
-    }
-
-    public Result Handle(EditPersonalInfoCommand command)
-    {
-      var unitOfWork = new UnitOfWork(_sessionFactory);
-      var repository = new StudentRepository(unitOfWork);
-      Student student = repository.GetById(command.Id);
-
-      if (student == null)
-        return Result.Fail($"No student found for Id {command.Id}");
-
-      student.Name = command.Name;
-      student.Email = command.Email;
-
-      unitOfWork.Commit();
-
-      return Result.Ok();
-    }
-  }
-  #endregion
-
-  //##################################################################### QUERY-BELOW #########################################
-  #region GetListQuery
-  public interface IQuery<TResult>
-  {
-
-  }
-  public interface IQueryHandler<TQuery, TResult>
-  where TQuery : IQuery<TResult>
-  {
-    TResult Handle(TQuery query);
-  }
-
-  public sealed class GetListQuery : IQuery<List<StudentDto>>
-  {
-    public string EnrolledIn { get; }
-    public int? NumberOfCourses { get; }
-
-    public GetListQuery(string enrolledIn, int? numberOfCourses)
-    {
-      EnrolledIn = enrolledIn;
-      NumberOfCourses = numberOfCourses;
-    }
-  }
-
-  public sealed class GetListQueryHandler : IQueryHandler<GetListQuery, List<StudentDto>>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public GetListQueryHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public List<StudentDto> Handle(GetListQuery query)
-    {
-      return new StudentRepository(_unitOfWork)
-                            .GetList(query.EnrolledIn, query.NumberOfCourses)
-                            .Select(x => ConvertToDto(x)).ToList();
-    }
-
-    private StudentDto ConvertToDto(Student student)
-    {
-      return new StudentDto
-      {
-        Id = student.Id,
-        Name = student.Name,
-        Email = student.Email,
-        Course1 = student.FirstEnrollment?.Course?.Name,
-        Course1Grade = student.FirstEnrollment?.Grade.ToString(),
-        Course1Credits = student.FirstEnrollment?.Course?.Credits,
-        Course2 = student.SecondEnrollment?.Course?.Name,
-        Course2Grade = student.SecondEnrollment?.Grade.ToString(),
-        Course2Credits = student.SecondEnrollment?.Course?.Credits,
-      };
-    }
-  } 
-  #endregion
-
-  //##################################################################### REMAINING-BELOW #########################################
-
-  #region RegisterCommand
-  public sealed class RegisterCommand : ICommand
-  {
-    public string Name { get; }
-    public string Email { get; }
-    public string Course1 { get; }
-    public string Course1Grade { get; }
-    public string Course2 { get; }
-    public string Course2Grade { get; }
-
-    public RegisterCommand(string name, string email, string course1, string course1Grade,
-      string course2, string course2Grade)
-    {
-      Name = name;
-      Email = email;
-      Course1 = course1;
-      Course1Grade = course1Grade;
-      Course2 = course2;
-      Course2Grade = course2Grade;
-    }
-  }
-
-  public sealed class RegisterCommandHandler : ICommandHandler<RegisterCommand>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public RegisterCommandHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public Result Handle(RegisterCommand command)
-    {
-      var courseRepository = new CourseRepository(_unitOfWork);
-      var studentRepository = new StudentRepository(_unitOfWork);
-      var student = new Student(command.Name, command.Email);
-
-      if (command.Course1 != null && command.Course1Grade != null)
-      {
-        Course course = courseRepository.GetByName(command.Course1);
-        student.Enroll(course, Enum.Parse<Grade>(command.Course1Grade));
-      }
-
-      if (command.Course2 != null && command.Course2Grade != null)
-      {
-        Course course = courseRepository.GetByName(command.Course2);
-        student.Enroll(course, Enum.Parse<Grade>(command.Course2Grade));
-      }
-
-      studentRepository.Save(student);
-      _unitOfWork.Commit();
-
-      return Result.Ok();
-    }
-
-  }
-  #endregion
-
-
-  #region UnregisterCommand
-  public sealed class UnregisterCommand : ICommand
-  {
-    public long Id { get; }
-
-    public UnregisterCommand(long id)
-    {
-      Id = id;
-    }
-  }
-
-  public sealed class UnregisterCommandHandler : ICommandHandler<UnregisterCommand>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public UnregisterCommandHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public Result Handle(UnregisterCommand command)
-    {
-      var studentRepository = new StudentRepository(_unitOfWork);
-      Student student = studentRepository.GetById(command.Id);
-      if (student == null)
-        return Result.Fail($"No student found for Id {command.Id}");
-
-      studentRepository.Delete(student);
-      _unitOfWork.Commit();
-
-      return Result.Ok();
-    }
-
-  }
-  #endregion
-
-  #region EnrollCommand
-  public sealed class EnrollCommand : ICommand
-  {
-    public long Id { get; }
-    public string Course { get; }
-    public string Grade { get; }
-
-    public EnrollCommand(long id, string course, string grade)
-    {
-      Id = id;
-      Course = course;
-      Grade = grade;
-    }
-  }
-
-  public sealed class EnrollCommandHandler : ICommandHandler<EnrollCommand>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public EnrollCommandHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public Result Handle(EnrollCommand command)
-    {
-      var studentRepository = new StudentRepository(_unitOfWork);
-      var courseRepository = new CourseRepository(_unitOfWork);
-
-      Student student = studentRepository.GetById(command.Id);
-      if (student == null)
-        return Result.Fail($"No student found for Id { command.Id }");
-
-      Course course = courseRepository.GetByName(command.Course);
-      if (course == null)
-        return Result.Fail($"Course is incorrect: '{ command.Course }'");
-
-      bool success = Enum.TryParse(command.Grade, out Grade grade);
-      if (!success)
-        return Result.Fail($"Grade is incorrect: '{ command.Grade }'");
-
-      student.Enroll(course, grade);
-      _unitOfWork.Commit();
-      return Result.Ok();
-    }
-
-  }
-  #endregion
-
-  #region TransferCommand
-  public sealed class TransferCommand : ICommand
-  {
-    public long Id { get; }
-    public int EnrollmentNumber { get; }
-    public string Course { get; }
-    public string Grade { get; }
-
-    public TransferCommand(long id, int enrollmentNumber, string course, string grade)
-    {
-      Id = id;
-      EnrollmentNumber = enrollmentNumber;
-      Course = course;
-      Grade = grade;
-    }
-  }
-
-  public sealed class TransferCommandHandler : ICommandHandler<TransferCommand>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public TransferCommandHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public Result Handle(TransferCommand command)
-    {
-      var studentRepository = new StudentRepository(_unitOfWork);
-      var courseRepository = new CourseRepository(_unitOfWork);
-
-      Student student = studentRepository.GetById(command.Id);
-      if (student == null)
-        return Result.Fail($"No student found for Id { command.Id }");
-
-      Course course = courseRepository.GetByName(command.Course);
-      if (course == null)
-        return Result.Fail($"Course is incorrect: '{ command.Course }'");
-
-      bool success = Enum.TryParse(command.Grade, out Grade grade);
-      if (!success)
-        return Result.Fail($"Grade is incorrect: '{ command.Grade }'");
-
-      Enrollment enrollment = student.GetEnrollment(command.EnrollmentNumber);
-      if (enrollment == null)
-        return Result.Fail($"No enrollment found with number ' { command.EnrollmentNumber } '");
-
-      enrollment.Update(course, enrollment.Grade);
-      _unitOfWork.Commit();
-
-      return Result.Ok();
-    }
-
-  }
-  #endregion
-
-  #region DisenrollCommand
-  public sealed class DisenrollCommand : ICommand
-  {
-    public long Id { get; }
-    public int EnrollmentNumber { get; }
-    public string Comment { get; }
-
-    public DisenrollCommand(long id, int enrollmentNumber, string comment)
-    {
-      Id = id;
-      EnrollmentNumber = enrollmentNumber;
-      Comment = comment;
-    }
-  }
-
-  public sealed class DisenrollCommandHandler : ICommandHandler<DisenrollCommand>
-  {
-    private readonly UnitOfWork _unitOfWork;
-
-    public DisenrollCommandHandler(UnitOfWork unitOfWork)
-    {
-      _unitOfWork = unitOfWork;
-    }
-
-    public Result Handle(DisenrollCommand command)
-    {
-      var studentRepository = new StudentRepository(_unitOfWork);
-
-      Student student = studentRepository.GetById(command.Id);
-      if (student == null)
-        return Result.Fail($"No student found for Id {command.Id}");
-
-      if (string.IsNullOrWhiteSpace(command.Comment))
-        return Result.Fail("Disenrollment comment is required");
-
-      Enrollment enrollment = student.GetEnrollment(command.EnrollmentNumber);
-      if (enrollment == null)
-        return Result.Fail($"No enrollment found with number ' { command.EnrollmentNumber } '");
-
-      student.RemoveEnrollment(enrollment, command.Comment);
-
-      _unitOfWork.Commit();
-
-      return Result.Ok();
-    }
-
-  } 
-  #endregion
-
-}
